requirements.txt
six==1.16.0
opencv-python==4.10.0.84
cvlib==0.2.7
gtts==2.5.4
ultralytics==8.3.31
flask==3.1.0
prisma==0.15.0
flasgger==0.9.7.1
python-dotenv==1.0.1
boto3==1.35.61
flask-socketio==5.4.1
eventlet==0.38.0
python-socketio==5.11.4
rembg==2.0.63
onnxruntime==1.20.1